Extract smoldot-related bindings in bindings-smoldot-js
